Remove cached revocation identifiers from `UnverifiedBiscuit`.
This simplifies the code and doesn't do unnecessary recomputation of the revoked ids collection
every time a new block is added. Likely, a relying party would only access the collection of revoked
ids once to check revocation anyway. This is in line with the rust implemention that also assembles
the collection of revoked identifiers on the fly.